requirements.txt
aniso8601==9.0.1
cairocffi==1.2.0
CairoSVG==2.5.2
cffi==1.14.5
click==8.0.1
colorama==0.4.4
cssselect2==0.4.1
defusedxml==0.7.1
Flask==2.0.1
Flask-RESTful==0.3.9
greenlet==1.1.0
html5lib==1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
protobuf==3.17.3
pycparser==2.20
pyphen==0.11.0
pytz==2021.1
six==1.16.0
tinycss2==1.1.0
WeasyPrint==52.5
webencodings==0.5.1
Werkzeug==2.0.1
